Unify main macros
