+++ compile-tests/wifi_1-1/src/main.rs
@@ -43,7 +43,7 @@ macro_rules! mk_static {
 const SSID: &str = "SSID";
 const PASSWORD: &str = "PASSWORD";
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
