+++ esp-hal-procmacros/src/lib.rs
@@ -22,20 +22,27 @@
 //!    sections, such as SRAM or RTC RAM (slow or fast) with different initialization options. See
 //!    its documentation for details.
 //!
-//!  - [`esp_rtos::main`](macro@rtos_main) - Creates a new instance of `esp_rtos::embassy::Executor`
-//!    and declares an application entry point spawning the corresponding function body as an async
-//!    task.
+//!  - [`main`](macro@main) - A unified entry point macro. For blocking functions it sets the
+//!    bare-metal entry point; for async functions it creates an `esp_rtos::embassy::Executor` and
+//!    spawns the function body as a task.
 //!
 //! ## Examples
 //!
-//! #### `main` macro
+//! #### Blocking entry point
 //!
-//! Requires the `embassy` feature to be enabled.
+//! ```rust,ignore
+//! #[main]
+//! fn main() -> ! {
+//!     loop { /* .. */ }
+//! }
+//! ```
+//!
+//! #### Async entry point (requires the `embassy` feature in `esp-rtos`)
 //!
 //! ```rust,ignore
 //! #[main]
 //! async fn main(spawner: Spawner) {
-//!     // Your application's entry point
+//!     // Your application's async entry point
 //! }
 //! ```
 //!
@@ -46,7 +53,6 @@
 use proc_macro::TokenStream;
 
 mod alert;
-mod blocking;
 mod builder;
 mod doc_replace;
 mod interrupt;
@@ -58,7 +64,7 @@ mod interrupt;
 ))]
 mod lp_core;
 mod ram;
-mod rtos_main;
+mod unified_main;
 
 /// Sets which segment of RAM to use for a function or static and how it should
 /// be initialized.
@@ -197,48 +203,17 @@ pub fn entry(args: TokenStream, input: TokenStream) -> TokenStream {
     lp_core::entry(args.into(), input.into()).into()
 }
 
-/// Creates a new instance of `esp_rtos::embassy::Executor` and declares an application entry point
-/// spawning the corresponding function body as an async task.
+/// Attribute to declare the entry point of the program.
 ///
-/// The following restrictions apply:
+/// Accepts both **blocking** and **async** main functions:
 ///
-/// * The function must accept exactly 1 parameter, an `embassy_executor::Spawner` handle that it
-///   can use to spawn additional tasks.
-/// * The function must be declared `async`.
-/// * The function must not use generics.
-/// * Only a single `main` task may be declared.
+/// - A **blocking** function (`fn main() -> !`) is treated as a bare-metal entry point.
+/// - An **async** function (`async fn main(spawner: Spawner)`) is spawned inside an
+///   `esp_rtos::embassy::Executor`.
 ///
 /// ## Examples
-/// Spawning a task:
 ///
-/// ```rust,ignore
-/// #[esp_rtos::main]
-/// async fn main(_s: embassy_executor::Spawner) {
-///     // Function body
-/// }
-/// ```
-#[proc_macro_attribute]
-pub fn rtos_main(args: TokenStream, item: TokenStream) -> TokenStream {
-    rtos_main::main(args.into(), item.into()).into()
-}
-
-/// Attribute to declare the entry point of the program
-///
-/// The specified function will be called by the reset handler *after* RAM has
-/// been initialized. If present, the FPU will also be enabled before the
-/// function is called.
-///
-/// The type of the specified function must be `[unsafe] fn() -> !` (never
-/// ending function)
-///
-/// # Properties
-///
-/// The entry point will be called by the reset handler. The program can't
-/// reference to the entry point, much less invoke it.
-///
-/// # Examples
-///
-/// - Simple entry point
+/// ### Blocking (bare-metal)
 ///
 /// ```ignore
 /// #[main]
@@ -246,10 +221,18 @@ pub fn rtos_main(args: TokenStream, item: TokenStream) -> TokenStream {
 ///     loop { /* .. */ }
 /// }
 /// ```
+///
+/// ### Async (embassy executor)
+///
+/// ```ignore
+/// #[main]
+/// async fn main(spawner: Spawner) {
+///     // spawn tasks, await futures, …
+/// }
+/// ```
 #[proc_macro_attribute]
-pub fn blocking_main(args: TokenStream, input: TokenStream) -> TokenStream {
-    let f = syn::parse_macro_input!(input as syn::ItemFn);
-    blocking::main(args.into(), f).into()
+pub fn main(args: TokenStream, input: TokenStream) -> TokenStream {
+    unified_main::main(args.into(), input.into()).into()
 }
 
 /// Automatically implement the [Builder Lite] pattern for a struct.
