+++ examples/async/embassy_hello_world/src/main.rs
@@ -21,7 +21,7 @@ async fn run() {
     }
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) {
     esp_println::logger::init_logger_from_env();
     let peripherals = esp_hal::init(esp_hal::Config::default());
