+++ examples/async/embassy_rmt_tx/src/main.rs
@@ -22,7 +22,7 @@ use esp_println::println;
 
 esp_bootloader_esp_idf::esp_app_desc!();
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(_spawner: Spawner) {
     println!("Init!");
