+++ examples/async/embassy_spi/src/main.rs
@@ -32,7 +32,7 @@ use esp_hal::{
 
 esp_bootloader_esp_idf::esp_app_desc!();
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(_spawner: Spawner) {
     esp_println::println!("Init!");
