+++ examples/async/embassy_usb_ethernet/src/main.rs
@@ -55,7 +55,7 @@ const HTTP_PORT: u16 = 80;
 
 esp_bootloader_esp_idf::esp_app_desc!();
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let peripherals = esp_hal::init(esp_hal::Config::default());
