+++ examples/esp-now/embassy_esp_now_duplex/src/main.rs
@@ -37,7 +37,7 @@ macro_rules! mk_static {
     }};
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
