+++ examples/wifi/embassy_access_point_with_sta/src/main.rs
@@ -74,7 +74,7 @@ macro_rules! mk_static {
     }};
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
