+++ examples/wifi/embassy_coex/src/main.rs
@@ -81,7 +81,7 @@ struct BatteryService {
     status: bool,
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
