+++ examples/wifi/embassy_sntp/src/main.rs
@@ -82,7 +82,7 @@ impl NtpTimestampGenerator for Timestamp<'_> {
     }
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) -> ! {
     esp_println::logger::init_logger_from_env();
     let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
