+++ qa-test/src/bin/embassy_adc_async_abstraction.rs
@@ -79,7 +79,7 @@ impl Converter for Identity {
     }
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(_spawner: Spawner) {
     esp_println::logger::init_logger_from_env();
     let peripherals = esp_hal::init(esp_hal::Config::default());
