+++ qa-test/src/bin/embassy_executor_benchmark.rs
@@ -79,7 +79,7 @@ async fn task3() {
     }
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) {
     esp_println::logger::init_logger_from_env();
