+++ qa-test/src/bin/embassy_wifi_i2s.rs
@@ -159,7 +159,7 @@ async fn i2s_dma_drain(
     }
 }
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) {
     // Enable logging from the ESP_LOG environment variable (set at build time)
     // Example: ESP_LOG=warn,esp_rtos=trace,esp_radio=info
