+++ qa-test/src/bin/embassy_wifi_stress.rs
@@ -34,7 +34,7 @@ esp_bootloader_esp_idf::esp_app_desc!();
 const SSID: &str = env!("SSID");
 const PASSWORD: &str = env!("PASSWORD");
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(_spawner: Spawner) {
     esp_println::logger::init_logger_from_env();
     let mut peripherals = initialize_esp_hal(EspConfig::default().with_cpu_clock(CpuClock::max()));
