+++ qa-test/src/bin/gpio_interrupt_latency.rs
@@ -32,7 +32,7 @@ esp_bootloader_esp_idf::esp_app_desc!();
 
 static SIGNAL: Signal<CriticalSectionRawMutex, ()> = Signal::new();
 
-#[esp_rtos::main]
+#[esp_hal::main]
 async fn main(spawner: Spawner) {
     esp_println::logger::init_logger_from_env();
     let p = esp_hal::init(esp_hal::Config::default().with_cpu_clock(CpuClock::max()));
